refactor: move models out of public directory
