+++ .gitattributes
@@ -1,1 +1,1 @@
-public/models/Xenova/paraphrase-multilingual-MiniLM-L12-v2/onnx/model_uint8.onnx filter=lfs diff=lfs merge=lfs -text
+models/Xenova/paraphrase-multilingual-MiniLM-L12-v2/onnx/model_uint8.onnx filter=lfs diff=lfs merge=lfs -text
