+++ models/Xenova/paraphrase-multilingual-MiniLM-L12-v2/config.json
