+++ models/Xenova/paraphrase-multilingual-MiniLM-L12-v2/onnx/model_uint8.onnx
